Split up `proto`
